Push constructor arguments into `InstanceCreator`
I think this is correct, but I'm not 100% sure of it. I've given the
`InstanceCreator` the means to resolve constructor arguments. I can't
explain at the moment why I think this might be a good or bad thing; I've
just got a niggle that it might not be a good refactoring.