Remove decorator pattern
Not for the first time in this code base, I find the decorator pattern
doesn't apply as cleanly as I'd thought. Returning anything from
`Execute()` is wrong, as we're only interested in the query results for
the purposes of throwing an exception when some are found. It's more of a
reporter than a query, so remove in the interface and rename the method,
and move it to the `Activation` namespace as it better fits that part of
the process.