Make DynamicBuilder's constructor internal
This requires deleting a test, but I think that is ok: the test was not
going via the public interface, and so was testing a scenario that won't
happen in real life. In order to encounter this scenario, a user would
have to somehow prevent the Named Arguments syntax parser from being run,
when the use of that parser is compiled into the assembly.