Delete the TODO list
All items except the last have been addressed. The final item - that
`MissingArgumentsQuery` and `PropertyValuesQuery` are the identical - is
not quite true: there's a pesky NOT operation in `MissingArgumentsQuery`.
I have a further refactoring in mind to completely encapsulate the
argument store's backing collection and operate on it only via a method
accepting an `IArgumentStoreQuery`, but that's for another branch.