Invoke Builders passed to builder methods
When a Builder is passed to a builder method, it needs to be invoked to
obtain the intended argument. This is done from the new
`BuilderArgumentEvaluator` type.

This requires the creation of a non-generic `DynamicBuilder` type to allow
`Build()` to be invoked without any compile-time knowledge of the type to
be constructed.